requirements.txt
tensorflow-gpu==1.13.1
scikit-image
matplotlib
numpy
